[FIX] ec.edu.espe.arquitectura.escolastico: Modificar Tipos

+++ src/main/java/ec/edu/espe/arquitectura/escolastico/educacion/dao/PeriodoRepository.java
@@ -5,7 +5,7 @@ import java.util.Date;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeriodoRepository extends JpaRepository<Periodo, Short> {
+public interface PeriodoRepository extends JpaRepository<Periodo, Integer> {
 
     List<Periodo> findByNombreLikeAndNivel(String nombrePattern, String nivel);
