+++ src/main/java/ec/edu/espe/arquitectura/escolastico/orgfisico/dao/AulaRepository.java
@@ -14,10 +14,10 @@ public interface AulaRepository extends JpaRepository<Aula, Integer> {
 
     List<Aula> findByCodEdificioBloque(String codEdificioBloque);
 
-    List<Aula> findByCodTipoAulaAndCapacidadGreaterThanOrderByCapacidadDesc(String codTipoAula, short capacidad);
+    List<Aula> findByCodTipoAulaAndCapacidadGreaterThanOrderByCapacidadDesc(String codTipoAula, Integer capacidad);
 
-    List<Aula> findByCodEdificioAndCapacidadGreaterThanOrderByCapacidadDesc(String codEdificio, short capacidad);
+    List<Aula> findByCodEdificioAndCapacidadGreaterThanOrderByCapacidadDesc(String codEdificio, Integer capacidad);
 
-    List<Aula> findByCodEdificioBloqueAndCapacidadGreaterThanOrderByCapacidadDesc(String codEdificioBloque, short capacidad);
+    List<Aula> findByCodEdificioBloqueAndCapacidadGreaterThanOrderByCapacidadDesc(String codEdificioBloque, Integer capacidad);
 
 }
