+++ src/main/java/ec/edu/espe/arquitectura/escolastico/orgfisico/dao/InstitucionRepository.java
@@ -3,7 +3,7 @@ package ec.edu.espe.arquitectura.escolastico.orgfisico.dao;
 import ec.edu.espe.arquitectura.escolastico.orgfisico.model.Institucion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstitucionRepository extends JpaRepository<Institucion, Short> {
+public interface InstitucionRepository extends JpaRepository<Institucion, Integer> {
 
     Institucion findByRuc(String ruc);
 }
